+++ testdata/config_modules.yml
@@ -42,9 +42,8 @@ mod_muc:
     allow_invites: false
     allow_subject_change: true
     enable_logging: true
-    history_length: 20
+    non_anonymous: true
     occupant_count: -1 # -1 means don't set the limit
     # options for the next ones are "all", "moderators" and ""
     real_jid_discovery: "all"
-    send_private_messages: "all"
-    can_get_member_list: "all"
+    send_pm: "all"
